requirements.txt
watchdog==6.0.0
google-api-python-client==2.100.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-auth==2.27.0
python-dateutil==2.8.2
setproctitle
